main.conf: document the new EnableIPv4LL option
When EnableIPv4LL is set to true, a link-local address from the IPv4 address
block 169.254.0.0/16 is assigned in case that no address can be obtained using
DHCP. The default is true.

This is not a new feature, but the ability to disable it has been added and
this commit documents the new option in the connman.conf(5) man page and the
sample main.conf.